Feat/javalin developer (#28)
* early setup for javalin

* dto

* get developer by email

* maybe adapter

* improve exception handler

* controller for developer hexagon

* modify DTO project skillstack name

* little change for javalin app

* setup postgre driver and spring jpa

* modify someshit

* added entities for database

* added getAllDeveloper to controller

* added mysql script and db entities

* feat: delete dev (not working)

* fix schema.sql

* add: return json on delete

* PostgreTest for testing

* add: adminer container

* add: get developer by skill route

* add: get developer by skill and experience

* add: create developer route

* fix: throw error

* Delete persistence.xml

* caca hibernate

---------

+++ src/main/java/domain/developers/DeveloperManager.java
@@ -5,6 +5,7 @@ import shared.Skill;
 import shared.developers.Email;
 import shared.exceptions.EntityAlreadyExistsException;
 import shared.exceptions.EntityNotFoundException;
+import shared.exceptions.InvalidAttributeException;
 import shared.projects.Name;
 
 
@@ -28,8 +29,8 @@ public class DeveloperManager implements ManageDeveloper, ManageDeveloperProject
 
 
     @Override
-    public Developer createDeveloper(Developer developer) throws EntityAlreadyExistsException {
-        if(developerExist(developer.getEmail())) {
+    public Developer createDeveloper(Developer developer) throws EntityAlreadyExistsException, InvalidAttributeException {
+        if(developerExist(new Email(developer.getEmailAddress()))) {
             throw new EntityAlreadyExistsException("The email provided is already used by a developer");
         }
 
@@ -53,6 +54,9 @@ public class DeveloperManager implements ManageDeveloper, ManageDeveloperProject
 
     @Override
     public Email removeDeveloper(Email email) {
+        if(!developerExist(email)) {
+            throw new EntityNotFoundException("The developer with the email provided does not exist");
+        }
         this.repository.removeDeveloper(email);
         return email;
     }
